+++ src/routes/(app)/+layout.svelte
@@ -1,17 +1,26 @@
 <script lang="ts">
   import type { PageData } from './$types';
   import { updated } from '$app/stores';
-  import Header from './_header.svelte';
+  import { clearAuthUser } from '$lib/stores/auth-user';
+  import { goto } from '$app/navigation';
+  import { routeForLoginPage } from '$lib/utilities/route-for';
+
+  import SideNavigation from './_side-nav.svelte';
   import Banners from '$lib/components/banner/banners.svelte';
   import { ErrorBoundary } from '$lib/holocene/error-boundary';
   import ScrollToTop from '$lib/holocene/scroll-to-top.svelte';
   import Toaster from '$lib/holocene/toaster.svelte';
   import { toaster } from '$lib/stores/toaster';
-  import TopActionNav from '$lib/components/top-action-nav.svelte';
+  import TopNavigation from '$lib/components/top-nav.svelte';
 
   export let data: PageData;
 
-  $: ({ user, uiVersionInfo } = data);
+  $: ({ uiVersionInfo } = data);
+
+  const logout = () => {
+    clearAuthUser();
+    goto(routeForLoginPage());
+  };
 
   updated.subscribe(async (value) => {
     if (value) {
@@ -24,10 +33,10 @@
 <div class="flex w-screen flex-row">
   <Toaster pop={toaster.pop} toasts={toaster.toasts} />
   <div class="sticky top-0 z-20 h-screen w-auto">
-    <Header {user} />
+    <SideNavigation />
   </div>
   <main id="content" class="min-h-screen w-max flex-auto bg-gray-100">
-    <TopActionNav />
+    <TopNavigation {logout} />
     <Banners {uiVersionInfo} />
     <div class="z-10 flex flex-col gap-4 px-10 pb-10 pt-8">
       <ErrorBoundary onError={() => {}}>
